// src/app/utils/seedSuperAdmin.ts
import { envVars } from "../../config/env"
import { IAuthProvider, IUser, Role } from "../modules/user/user.interface";
import { User } from "../modules/user/user.model"
import bcryptjs from 'bcryptjs';

export const seedSuperAdmin = async () => {
    try {
        const isSuperAdminExist = await User.findOne({ email: envVars.SUPER_ADMIN_EMAIL });
        if (isSuperAdminExist) {
            console.log("Is Super Admin already exist")
            return
        };

        console.log("Create trying super admin...")

        const hashPassword = await bcryptjs.hash(envVars.SUPER_ADMIN_EMAIL, Number(envVars.BCRYPT_SALT_ROUND));

        const authProvider: IAuthProvider = {
            provider: "credentials",
            providerId: envVars.SUPER_ADMIN_EMAIL
        };

        const payload: IUser = {
            name: "Super Admin",
            role: Role.SUPER_ADMIN,
            email: envVars.SUPER_ADMIN_EMAIL,
            password: hashPassword,
            isVerified: true,
            auths: [authProvider]
        };


        const superAdmin = await User.create(payload);
        console.log(superAdmin);
        console.log("Super Admin created Successful");
    } catch (error) {
        console.log(error);
    }
}